+++ NEWS.md
@@ -1,6 +1,7 @@
 vivo 0.1.1
 ----------------------------------------------------------------
 * fix in `calculate_weight()`, change in `ingredients` package
+* add function `calculate_variable_split()`
 
 vivo 0.1.0
 ----------------------------------------------------------------